mesa: Upgrade to 19.2.3
Upgrade mesa and mesa-gl to 19.2.3.

The license hash change was a trivial new line removal.

The glx-tls option was removed as it isn't included in the meson.build
file. It has been replaced with 'use-elf-tls' instead. We also added an
INSANE_SKIP to fix this error
    do_package_qa: QA Issue: ELF binary
    '[...]libgles2-mesa/usr/lib/libGLESv2.so.2.0.0' has relocations in .text
    [textrel]

The -Dasm=false was removed as it also is no longer included.

Signed-off-by: Alistair Francis <alistair@alistair23.me>
Signed-off-by: Alistair Francis <alistair.francis@wdc.com>
Acked-by: Otavio Salvador <otavio@ossystems.com.br>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>